Fix 'envisatdataset.cpp(136) : warning C4267: '+=' : conversion from 'size_t' to 'unsigned int', possible loss of data'
git-svn-id: https://svn.osgeo.org/gdal/trunk@31698 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/envisat/envisatdataset.cpp
@@ -133,7 +133,7 @@ CPLErr MerisL2FlagBand::IReadBlock( CPL_UNUSED int nBlockXOff,
 
     for( unsigned iImg = 0, iBuf = 0;
          iImg < nBlockXSize * (unsigned)sizeof(GDT_UInt32);
-         iImg += (unsigned)sizeof(GDT_UInt32), iBuf += nBytePerPixel )
+         iImg += (unsigned)sizeof(GDT_UInt32), iBuf += (unsigned)nBytePerPixel )
     {
 #ifdef CPL_LSB
         ((GByte*) pImage)[iImg] = pReadBuf[iBuf + 2];
